refactor: rename purchase into subscriptions
Purchase was GitHub related but more and more our customers are on
Stripe. Our model need to be consistent with what we are doing. That's
why I change naming and use "subscription" instead of "purchase".

Additionnaly I changed "stripePlanId" to "stripeProductId" and
"githubId" to "githubPlanId" in "plan" table.

The previous "getSubscription" is not "getSubscriptionManager", it
allows to manage the subscription, so it makes sense to call it like
that.