fix: fix auth behaviour
The user can now attach GitHub, GitLab and Google to the same account
and manage it from settings.

When logged from Google, it will now be seemless to be able to import a
repository from GitHub.

Also added a login management UI in the personal settings of the user.

Fix #996