chore(deps): upgrade dependencies
Remove fragment masking. It is a bit useless with TypeScript and could slow down the app and the development. Also the new Apollo has the feature integraded that is incompatible with gql-gen. We may re-enable it at some point but for now it is useless for us, TypeScript provides enough security.