feat: aggregate test stats
We aggregate test stats into two tables: "test_stats_builds" and "test_stats_changes".
We store the number of reference builds a test is present in and also
the number of changes by fileId a test has.
The computation will be done after each screenshot diff process (on a
reference build).
These two tables can be used to compute all stats without having to rely
on "screenshot_diffs" table.
It will allow us to create a test dashboard for a project.